fix CI/CD

+++ python/sglang/srt/lora/backend/triton_backend.py
@@ -197,7 +197,7 @@ class TritonLoRABackend(BaseLoRABackend):
                     (bs,), dtype=torch.int32, device=self.device
                 ),
                 lora_ranks=torch.empty(
-                    (self.max_loras_per_batch,), dtype=torch.int32, device=self.device
+                    (self.max_loras_per_batch,), dtype=torch.int64, device=self.device
                 ),
                 scalings=torch.empty(
                     (self.max_loras_per_batch,), dtype=torch.float, device=self.device
